refactor: move rule processor testing helpers to dedicated module
