+++ libs/methodologies/bold/rule-processors/mass-id/geolocation-precision/src/geolocation-precision.processor.spec.ts
@@ -1,8 +1,8 @@
+import { spyOnDocumentQueryServiceLoad } from '@carrot-fndn/shared/methodologies/bold/io-helpers';
 import {
   createRuleTestFixture,
   expectRuleOutput,
-  spyOnDocumentQueryServiceLoad,
-} from '@carrot-fndn/shared/methodologies/bold/io-helpers';
+} from '@carrot-fndn/shared/methodologies/bold/testing';
 import { type RuleInput } from '@carrot-fndn/shared/rule/types';
 import { random } from 'typia';
 
@@ -33,6 +33,7 @@ describe('GeolocationPrecisionProcessor', () => {
         massIdActorParticipants: actorParticipants,
         massIdDocumentsParams: massIdDocumentParameters,
         ruleDataProcessor,
+        spyOnDocumentQueryServiceLoad,
       });
 
       expectRuleOutput({
