+++ libs/methodologies/bold/rule-processors/mass-id/mass-id-sorting/src/mass-id-sorting.processor.spec.ts
@@ -1,8 +1,8 @@
+import { spyOnDocumentQueryServiceLoad } from '@carrot-fndn/shared/methodologies/bold/io-helpers';
 import {
   createRuleTestFixture,
   expectRuleOutput,
-  spyOnDocumentQueryServiceLoad,
-} from '@carrot-fndn/shared/methodologies/bold/io-helpers';
+} from '@carrot-fndn/shared/methodologies/bold/testing';
 import { type RuleInput } from '@carrot-fndn/shared/rule/types';
 import { random } from 'typia';
 
@@ -36,6 +36,7 @@ describe('MassIdSortingProcessor', () => {
             externalEventsMap: massIdEvents,
           },
           ruleDataProcessor,
+          spyOnDocumentQueryServiceLoad,
         });
 
         expectRuleOutput({
