+++ libs/methodologies/bold/rule-processors/mass-id/prevented-emissions/src/prevented-emissions.processor.spec.ts
@@ -1,8 +1,8 @@
+import { spyOnDocumentQueryServiceLoad } from '@carrot-fndn/shared/methodologies/bold/io-helpers';
 import {
   createRuleTestFixture,
   expectRuleOutput,
-  spyOnDocumentQueryServiceLoad,
-} from '@carrot-fndn/shared/methodologies/bold/io-helpers';
+} from '@carrot-fndn/shared/methodologies/bold/testing';
 import { type RuleInput } from '@carrot-fndn/shared/rule/types';
 import { random } from 'typia';
 
@@ -39,6 +39,7 @@ describe('PreventedEmissionsProcessor', () => {
             },
           },
           ruleDataProcessor,
+          spyOnDocumentQueryServiceLoad,
         });
 
         expectRuleOutput({
