+++ libs/methodologies/bold/rule-processors/mass-id/weighing/src/weighing.processor.spec.ts
@@ -1,8 +1,8 @@
+import { spyOnDocumentQueryServiceLoad } from '@carrot-fndn/shared/methodologies/bold/io-helpers';
 import {
   createRuleTestFixture,
   expectRuleOutput,
-  spyOnDocumentQueryServiceLoad,
-} from '@carrot-fndn/shared/methodologies/bold/io-helpers';
+} from '@carrot-fndn/shared/methodologies/bold/testing';
 import { type RuleInput } from '@carrot-fndn/shared/rule/types';
 import { random } from 'typia';
 
@@ -34,6 +34,7 @@ describe('WeighingProcessor', () => {
             externalEventsMap: massIdDocumentEvents,
           },
           ruleDataProcessor,
+          spyOnDocumentQueryServiceLoad,
         });
 
         expectRuleOutput({
