+++ libs/shared/methodologies/bold/io-helpers/src/testing.helpers.ts
@@ -1,14 +1,5 @@
 import type { Document } from '@carrot-fndn/shared/methodologies/bold/types';
 
-import { RuleDataProcessor } from '@carrot-fndn/shared/app/types';
-import {
-  BoldStubsBuilder,
-  type StubBoldDocumentParameters,
-} from '@carrot-fndn/shared/methodologies/bold/testing';
-import { RuleInput, RuleOutput } from '@carrot-fndn/shared/rule/types';
-import { AnyObject, MethodologyParticipant } from '@carrot-fndn/shared/types';
-import { random } from 'typia';
-
 import { DocumentQueryService } from './document-query.service';
 import * as documentHelpers from './document.helpers';
 
@@ -35,91 +26,3 @@ export const spyOnLoadDocument = (
 ) => {
   jest.spyOn(documentHelpers, 'loadDocument').mockResolvedValueOnce(result);
 };
-
-interface CreateBoldStubsParameters {
-  homologationDocuments?: Map<string, StubBoldDocumentParameters> | undefined;
-  massIdActorParticipants?: Map<string, MethodologyParticipant> | undefined;
-  massIdDocumentsParams?: StubBoldDocumentParameters | undefined;
-}
-
-interface ProcessRuleTestParameters {
-  homologationDocuments?: Map<string, StubBoldDocumentParameters> | undefined;
-  massIdActorParticipants?: Map<string, MethodologyParticipant> | undefined;
-  massIdDocumentsParams?: StubBoldDocumentParameters | undefined;
-  ruleDataProcessor: RuleDataProcessor;
-}
-
-interface TestExpectedRuleOutputParameters {
-  resultComment: string;
-  resultContent?: AnyObject | undefined;
-  resultStatus: string;
-  ruleInput: RuleInput;
-  ruleOutput: RuleOutput;
-}
-
-export function createBoldStubsForMassIdProcessor({
-  homologationDocuments,
-  massIdActorParticipants,
-  massIdDocumentsParams,
-}: CreateBoldStubsParameters) {
-  return new BoldStubsBuilder({
-    ...(massIdActorParticipants && { massIdActorParticipants }),
-  })
-    .createMassIdDocuments(massIdDocumentsParams)
-    .createMassIdAuditDocuments()
-    .createMethodologyDocument()
-    .createParticipantHomologationDocuments(homologationDocuments)
-    .build();
-}
-
-export async function createRuleTestFixture({
-  homologationDocuments,
-  massIdActorParticipants,
-  massIdDocumentsParams,
-  ruleDataProcessor,
-}: ProcessRuleTestParameters): Promise<{
-  ruleInput: RuleInput;
-  ruleOutput: RuleOutput;
-}> {
-  const {
-    massIdAuditDocument,
-    massIdDocument,
-    participantsHomologationDocuments,
-  } = createBoldStubsForMassIdProcessor({
-    homologationDocuments,
-    massIdActorParticipants,
-    massIdDocumentsParams,
-  });
-
-  const allDocuments = [
-    massIdDocument,
-    massIdAuditDocument,
-    ...participantsHomologationDocuments.values(),
-  ];
-
-  spyOnDocumentQueryServiceLoad(massIdAuditDocument, allDocuments);
-
-  const ruleInput = {
-    ...random<Required<RuleInput>>(),
-    documentId: massIdAuditDocument.id,
-  };
-
-  return { ruleInput, ruleOutput: await ruleDataProcessor.process(ruleInput) };
-}
-
-export function expectRuleOutput({
-  resultComment,
-  resultContent,
-  resultStatus,
-  ruleInput,
-  ruleOutput,
-}: TestExpectedRuleOutputParameters): void {
-  expect(ruleOutput).toEqual({
-    requestId: ruleInput.requestId,
-    responseToken: ruleInput.responseToken,
-    responseUrl: ruleInput.responseUrl,
-    resultComment,
-    resultStatus,
-    ...(resultContent !== undefined && { resultContent }),
-  });
-}
