+++ libs/shared/methodologies/bold/testing/src/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './rule-processor.helpers';
