+++ libs/shared/methodologies/bold/testing/src/index.ts
@@ -1,2 +1,3 @@
 export * from './builders';
+export * from './helpers';
 export * from './stubs';
